rename to Status & move to commands.go
