+++ tests/e2e/configurer/chain/node.go
@@ -2,7 +2,6 @@ package chain
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"regexp"
 	"strings"
@@ -13,11 +12,8 @@ import (
 	rpchttp "github.com/tendermint/tendermint/rpc/client/http"
 	coretypes "github.com/tendermint/tendermint/rpc/core/types"
 
-	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	"github.com/osmosis-labs/osmosis/v14/tests/e2e/containers"
 	"github.com/osmosis-labs/osmosis/v14/tests/e2e/initialization"
-	"github.com/tendermint/tendermint/libs/bytes"
-	"github.com/tendermint/tendermint/p2p"
 )
 
 type NodeConfig struct {
@@ -148,28 +144,3 @@ func (n *NodeConfig) LogActionF(msg string, args ...interface{}) {
 	n.t.Logf("[%s] %s. From container %s", timeSinceStart, s, n.Name)
 }
 
-func (n *NodeConfig) NodeStatus() error {
-	cmd := []string{"osmosisd", "status"}
-	_, errBuf, err := n.containerManager.ExecCmd(n.t, n.Name, cmd, "")
-	if err != nil {
-		return err
-	}
-
-	type validatorInfo struct {
-		Address     bytes.HexBytes
-		PubKey      cryptotypes.PubKey
-		VotingPower int64
-	}
-
-	type resultStatus struct {
-		NodeInfo      p2p.DefaultNodeInfo
-		SyncInfo      coretypes.SyncInfo
-		ValidatorInfo validatorInfo
-	}
-	var result resultStatus
-	err = json.Unmarshal(errBuf.Bytes(), &result)
-	if err != nil {
-		return err
-	}
-	return nil
-}
